# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the NutriMind nutrition and health tracking mobile app.",
        "title": "NutriMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/social/challenges": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all available challenges with their type, duration, description, how many of the user's friends are enrolled, and the user's own enrollment summary if they have joined.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Get challenge catalogue",
                "responses": {
                    "200": {
                        "description": "Challenge catalogue",
                        "schema": {
                            "$ref": "#/definitions/dto.GetChallengeCatalogueResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "SOCIAL_DISABLED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/social/challenges/{challenge_id}/enrollment": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks the user's enrollment in a challenge as abandoned. The user may re-join later.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Abandon a challenge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "challenge_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Enrollment abandoned"
                    },
                    "400": {
                        "description": "Invalid challenge_id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "SOCIAL_DISABLED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Active enrollment not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/social/challenges/{challenge_id}/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enrols the authenticated user in the specified challenge. Returns 409 if they are already enrolled in an active run of this challenge.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Join a challenge",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "challenge_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Enrollment created with start/end dates",
                        "schema": {
                            "$ref": "#/definitions/dto.JoinChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid challenge_id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "SOCIAL_DISABLED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Challenge not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already enrolled",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/social/challenges/{challenge_id}/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the day-grid progress for the current user and their enrolled friends for a given challenge. Each grid day indicates whether the user met their daily goal.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social"
                ],
                "summary": "Get challenge progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Challenge ID",
                        "name": "challenge_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progress for user and friends",
                        "schema": {
                            "$ref": "#/definitions/dto.GetChallengeProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid challenge_id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "SOCIAL_DISABLED",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Challenge or enrollment not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CatalogueChallengeItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration_days": {
                    "type": "integer"
                },
                "friends_enrolled": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "my_enrollment": {
                    "$ref": "#/definitions/dto.EnrollmentSummaryResponse"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.ChallengeInfoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetChallengeCatalogueResponse": {
            "type": "object",
            "properties": {
                "catalogue": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CatalogueChallengeItemResponse"
                    }
                }
            }
        },
        "dto.GetChallengeProgressResponse": {
            "type": "object",
            "properties": {
                "challenge": {
                    "$ref": "#/definitions/dto.ChallengeInfoResponse"
                },
                "friends_progress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FriendProgressResponse"
                    }
                },
                "my_progress": {
                    "$ref": "#/definitions/dto.MyProgressResponse"
                }
            }
        },
...
        "dto.JoinChallengeResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "enrollment_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.Error"
                },
                "is_success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token — format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}